refactor: widen the lint gate, delete dead weight, make main.rs testable
Three independent maintenance changes with no runtime behaviour change. CLI
help text for all 24 subcommands and every `*-shell-env` output were verified
byte-identical before and after.

`make lint` was checking 2 of the 5 workspace crates and no test targets at
all, so `amr-safe`, `amr-sys`, `pjsua-sys`, every integration test and every
`#[cfg(test)]` module went unlinted -- hiding ~15 warnings including entirely
dead test-support code. It is now `--workspace --all-targets -D warnings`, and
CLAUDE.md says not to narrow it again.

`deny.toml` was not merely dated, it was broken: `[advisories] vulnerability`
/`notice` and `[licenses] unlicensed` are *removed* keys that hard-error on
current cargo-deny, so `cargo deny check` failed outright for anyone who had
the tool -- hidden by `make lint`'s `if command -v` guard. Rewritten against
the current schema, and CI now installs cargo-deny (and cargo-nextest, for the
per-test timeout `.config/nextest.toml` describes but nothing ever ran) so
both are genuinely enforced somewhere.

`src/main.rs` held 40 CLI handlers in 2099 lines. Items in a binary crate
cannot be imported from `tests/`, so all of it -- line resolution, call
reporting, and the `*-shell-env` printers whose output docker/healthcheck.sh
`eval`s -- had zero tests for a purely structural reason. Moved verbatim into
`gsm_sip_bridge::commands` (one module per subcommand family, plus
`commands::daemon`); main.rs is now 28 lines. The 24-arm
`if let Some(Commands::X(..))` chain became a `match`, so a subcommand that is
not wired up is a build error rather than one that silently falls through and
starts the daemon. The three shell-env printers now return a String instead of
writing to stdout, which is what makes them assertable -- six new tests pin
the array-length invariant healthcheck.sh's `seq` loop depends on, the
zero-lines case, and shell_quote escaping.

Deleted: docker/grafana/dashboards/gsm-bridge.json (mounted nowhere --
docker-compose only mounts grafana/provisioning -- and all 28 panels query the
pre-v5 `gsm_bridge_*` metric names, so nothing was worth porting),
tests/common/ (fully unused, plus the 25 `mod common;` declarations that
existed only to satisfy it), and the no-op `make test-bash` target.

881 tests pass (875 + 6 new). `make lint` clean. Zero unsafe in the app crate.
