+++ CLAUDE.md
@@ -9,12 +9,14 @@ shell commands, and other important information, read the current plan at
 **MANDATORY — run before EVERY commit, no exceptions:**
 
 ```bash
-cargo fmt --all          # fix formatting in place
-make lint                # rustfmt check + clippy -D warnings + unsafe ratio
-cargo test --workspace   # all tests must pass
+make format              # fix formatting in place
+make lint                # clippy --workspace --all-targets -D warnings, + deny/shellcheck/unsafe
+make test                # all tests must pass
 ```
-
-Or equivalently: `make format && make lint && make test`
 
 Do NOT commit if any of these fail. `make lint` failing has caused broken
 commits in the past (e.g. rustfmt line-length violations in test files).
+
+`make lint` covers the **whole workspace including all test targets** — a
+warning in an integration test or a `#[cfg(test)]` module fails the build
+exactly like one in production code. Do not narrow its scope back down.
