+++ gsm-sip-bridge/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod alerts;
 pub mod cli;
+pub mod commands;
 pub mod config;
 pub mod control;
 pub mod error;
