+++ gsm-sip-bridge/tests/test_end_to_end.rs
@@ -1,5 +1,3 @@
-mod common;
-
 use gsm_sip_bridge::modules::audio_pipeline::AudioPipeline;
 use gsm_sip_bridge::modules::card::{CardInstance, CardState};
 use std::sync::Arc;
