+++ gsm-sip-bridge/tests/test_migration_sql.rs
@@ -1,5 +1,3 @@
-mod common;
-
 use rusqlite::Connection;
 use tempfile::NamedTempFile;
