+++ gsm-sip-bridge/tests/test_sms_reader.rs
@@ -1,6 +1,4 @@
-mod common;
-
-use gsm_sip_bridge::modules::at_commander::{AtCommander, AtResponse};
+use gsm_sip_bridge::modules::at_commander::AtCommander;
 use gsm_sip_bridge::sms::reader::{delete_sms, read_sms};
 use std::io::{Read, Write};
 use std::time::Duration;
