+++ gsm-sip-bridge/tests/test_vowifi_call_history.rs
@@ -1,5 +1,3 @@
-mod common;
-
 // User Story 3 (specs/014-vowifi-metrics-restore): every inbound VoWiFi
 // call must land in the shared `calls` table, the same one circuit-switched
 // calls use, with `transport='vowifi'`. Drives the real
